Split dogtab.test.js into two files
