+++ tests/dogtab.test.js
@@ -5,10 +5,7 @@ const got = require('got');
 
 const app = require('../index.js');
 
-const { dogtabsDogTabIDGET, dogtabsGET, dogtabsPOST, 
-    dogtabsSavedDogTabIDDELETE, dogtabsSavedDogTabIDPUT, 
-    dogtabsSavedGET, dogtabsSexBreedAgeRangeLocationGET } = require("../service/DogTabService.js")
-
+//create context
 test.before(async (t) => {
     t.context.server = http.createServer(app);
     t.context.prefixUrl = await listen(t.context.server);
@@ -17,19 +14,6 @@ test.before(async (t) => {
 
 test.after.always((t) => {
     t.context.server.close();
-});
-
-test("GET posted DogTab by ID by function", async t => {
-    // define parameters
-    const id = 1234567;
-    // get posted DogTab by ID
-    const result = await dogtabsDogTabIDGET(id);
-    // check some dogtab characteristics
-    t.is(result.sex, "Male");
-    t.is(result.name, "Rex");
-    t.is(result.description, "Rex seeks a home");
-    t.is(result.location, "Thessaloniki, Greece");
-    t.is(result.breed, "Bulldog");
 });
 
 test("GET posted DogTab by ID", async t => {
@@ -56,21 +40,6 @@ test("Assure GET posted DogTab by ID fails with non integer IDs", async t => {
     t.is(error.response.body.message, "request.params.dogTabID should be integer");
 });
 
-test("GET all posted DogTabs by function", async t => {
-    //get all dogtabs
-    const result = await dogtabsGET();
-    // assert that we get two entries 
-    t.is(result.length, 2);
-    // get the first one
-    const firstDogtab = result[0];
-    // check some dogtab characteristics
-    t.is(firstDogtab.sex, "Male");
-    t.is(firstDogtab.name, "Rex");
-    t.is(firstDogtab.description, "Rex seeks a home");
-    t.is(firstDogtab.location, "Thessaloniki, Greece");
-    t.is(firstDogtab.breed, "Bulldog");
-});
-
 test("GET all posted DogTabs", async t => {
     const { body, statusCode } = await t.context.got("dogtabs");
     // assert success status code
@@ -85,30 +54,6 @@ test("GET all posted DogTabs", async t => {
     t.is(firstDogtab.description, "Rex seeks a home");
     t.is(firstDogtab.location, "Thessaloniki, Greece");
     t.is(firstDogtab.breed, "Bulldog");
-});
-
-test("POST new DogTab by function", async t => {
-    // define parameters
-    const body = {
-        "name": "Rex",
-        "breed": "Bulldog",
-        "sex": "Male",
-        "mainPhoto": "",
-        "otherPhotos": [
-          ""
-        ],
-        "birthDate": "2023-12-12T11:37:48.618Z",
-        "description": "Rex seeks a home",
-        "location": "Thessaloniki, Greece"
-    };
-    // try to create dogtab
-    try {
-        await dogtabsPOST(body);
-        t.pass();
-    } catch (e) {
-        // catch error
-        t.fail(e);
-    }
 });
 
 test("POST new DogTab", async t => {
@@ -158,20 +103,6 @@ test("Assure POST new DogTab fails with wrong body", async t => {
     t.is(error.response.body.message, "request.body.name should be string, request.body.breed should be string");
 });
 
-test("DELETE DogTab from own interest list by ID by function", async t => {
-    // define parameters
-    const id = 1234567;
-    //try to delete DogTab from interest list
-    try {
-        await dogtabsSavedDogTabIDDELETE(id);
-        t.pass();
-    } 
-    // catch error
-    catch(e){
-        t.fail(e);
-    }
-});
-
 test("DELETE DogTab from own interest list by ID", async t => {
     // define parameters
     const id = 1234567;
@@ -189,20 +120,6 @@ test("Assure DELETE DogTab from own interest list by ID fails with non integer I
     });
     t.is(error.response.statusCode, 400);
     t.is(error.response.body.message, "request.params.dogTabID should be integer");
-});
-
-test("PUT DogTab into own interest list by ID by function", async t => {
-    // define parameters
-    const id = 1234567;
-    //try to add DogTab in interest list
-    try {
-        await dogtabsSavedDogTabIDPUT(id);
-        t.pass();
-    } 
-    // catch error
-    catch(e){
-        t.fail(e);
-    }
 });
 
 test("PUT DogTab into own interest list by ID", async t => {
@@ -224,21 +141,6 @@ test("Assure PUT DogTab into own interest list by ID fails with non integer IDs"
     t.is(error.response.body.message, "request.params.dogTabID should be integer");
 });
 
-test("GET all saved DogTabs by function", async t => {
-    // get all saved DogTabs
-    const result = await dogtabsSavedGET();
-    // assert that we get two entries 
-    t.is(result.length, 2);
-    // get the first one
-    const firstDogtab = result[0];
-    // check some dogtab characteristics
-    t.is(firstDogtab.sex, "Male");
-    t.is(firstDogtab.name, "Rex");
-    t.is(firstDogtab.description, "Rex seeks a home");
-    t.is(firstDogtab.location, "Thessaloniki, Greece");
-    t.is(firstDogtab.breed, "Bulldog");
-});
-
 test("GET all saved DogTabs", async t => {
     // make get request to mock server
     const { body, statusCode } = await t.context.got(`dogtabs/saved`);
@@ -254,24 +156,6 @@ test("GET all saved DogTabs", async t => {
     t.is(firstDogTab.description, "Rex seeks a home");
     t.is(firstDogTab.location, "Thessaloniki, Greece");
     t.is(firstDogTab.breed, "Bulldog");
-});
-
-test("GET DogTabs by filters by function", async t => {
-    // define parameters
-    const sex = "Male";
-    const breed = "Bulldog";
-    const ageRange = "1m0y - 0m10y";
-    const location = "Thessaloniki";
-    // get all filtered DogTabs
-    const result = await dogtabsSexBreedAgeRangeLocationGET(sex, breed, ageRange, location);
-    // assert that we get two entries 
-    t.is(result.length, 2);
-    // get the first one    
-    const firstDogtab = result[0];
-    // check some dogtab characteristics
-    t.is(firstDogtab.sex, "Male");
-    t.is(firstDogtab.location, "Thessaloniki, Greece");
-    t.is(firstDogtab.breed, "Bulldog");
 });
 
 test("GET DogTabs by filters", async t => {
